backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
flake8==7.3.0
gunicorn==24.1.1
idna==3.11
mccabe==0.7.0
oauthlib==3.3.1
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.2.1
python3-openid==3.2.0
reportlab==4.4.1
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
